Only show banner on shell requests
Without this we see the banner twice on `ssh example.com`, (one when requesting a pseudo terminal, and one when opening a shell session), and the connection hangs.

Things are marginally more sensible with `ssh -T example.com`, which disables pseudo terminal requests, but not much.

This change, then, instead will run on shell requests (which are the only request that pty/no-pty requests have in common) and nothing else